Rename __test_utils__ -> __utils__
